refactor: add classmate dependency to support generic types as registry keys
